Port of twitch_auth.rs file to separate lib.
The `src/lib.rs` file is a direct port of
[`twitch_auth.rs`](https://github.com/stuck-overflow/ferris-bot/blob/fcd3d380cf3adccaed486c97a0a4ca0268884bb1/src/twitch_auth.rs)

This library temporarily requires the twitch_api2 library to be
available as a "sister" directory (i.e. in `../twitch_api2`). This is
needed since our other projects are using unreleased features of
twitch_api2 and we need to build against the same codebase.